chore: extract DatabaseContext into its own crate
