+++ crates/api/Cargo.toml
@@ -21,5 +21,6 @@ util.workspace = true
 sd-notify = "0.4.1"
 
 [dev-dependencies]
+db-context.workspace = true
 tempfile.workspace = true
 tokio = "1.34.0"
