+++ crates/db-context/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "util"
+name = "db-context"
 version = "1.0.0"
 authors.workspace = true
 edition.workspace = true
@@ -10,10 +10,8 @@ publish = false  # TODO: figure out a better name for this crate before enabling
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-db-context.workspace = true
+camino = "1.1.6"
 futures = "0.3.29"
 sqlx = { version = "0.7.3", features = ["postgres", "runtime-tokio"] }
-tokio = { version = "1.34.0", features = ["process", "macros", "rt", "rt-multi-thread"] }
-
-[build-dependencies]
-camino = "1.1.6"
+tempfile = "3.8.1"
+fragile-child = { version = "1.0.0", path = "../fragile-child" }
