+++ crates/fetcher/Cargo.toml
@@ -26,5 +26,6 @@ sqlx-core = "0.7.3"
 serde_json = "1.0.111"
 
 [dev-dependencies]
+db-context.workspace = true
 once_cell = "1.19.0"
 tempfile.workspace = true
