+++ crates/fetcher/src/test.rs
@@ -1,9 +1,9 @@
 use camino::Utf8PathBuf;
+use db_context::DatabaseContext;
 use futures::{future::LocalBoxFuture, FutureExt};
 use itertools::Itertools;
 use pr_tracker_store::{Branch, GitCommit, GithubPrQueryCursor, Landing, Pr, PrNumber};
 use tempfile::TempDir;
-use util::DatabaseContext;
 use wildmatch::WildMatch;
 
 use crate::{github::GithubClient, repo::isolated_git};
