+++ crates/util/src/bin/db-repl.rs
@@ -1,6 +1,6 @@
+use db_context::DatabaseContext;
 use futures::FutureExt;
 use std::process::{self, Command};
-use util::DatabaseContext;
 
 #[tokio::main]
 async fn main() {
