+++ crates/util/src/bin/sqlx-prepare.rs
@@ -1,6 +1,6 @@
+use db_context::DatabaseContext;
 use futures::FutureExt;
 use std::process::Command;
-use util::DatabaseContext;
 
 #[tokio::main]
 async fn main() {
